requirements.txt
# ===============================
# Core Web App
# ===============================
flask==3.0.0

# ===============================
# Data & ML
# ===============================
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# ===============================
# RAG (Retrieval-Augmented Generation)
# ===============================
faiss-cpu==1.8.0
sentence-transformers==2.6.1

# ===============================
# LLM (Groq)
# ===============================
groq==0.9.0
httpx==0.27.2
httpcore==1.0.4

# ===============================
# Utilities
# ===============================
tqdm==4.66.1
